+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Added guild generation
 
+- Fixed all views to be justified
+
+- Made all views selectable with `SelectionArea` rather than `SelectableText`
+
 ## 0.7.0
 
 - Added deities generation